Move common bt-specific code into bt_common.h
This allows bt_producer.h to no longer include all of bt_serialize.h
just for the concepts they share.
